backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pgvector==0.2.5
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
pandas==2.2.2
openpyxl==3.1.2
pdfplumber==0.11.0
PyMuPDF==1.24.4
groq==0.8.0
google-generativeai>=0.8.0
xlsxwriter==3.2.0
matplotlib==3.9.0
seaborn==0.13.2
kaleido==0.2.1
plotly==5.22.0
numpy==1.26.4
python-dotenv==1.0.1
aiofiles==23.2.1
httpx==0.27.0
